+++ packages/evolution-generator/src/scripts/generate_survey.py
@@ -145,9 +145,18 @@ def generate_survey(config_path):
         # TODO: At some point, we should consider only read the Excel sheet one time, to avoid reading it multiple times.
         # TODO: We might consider extracting the sheet names from the Excel file or config file instead of hardcoding them.
         # Generate the labels for the specified sheets
-        sheet_names = ["Widgets", "Labels"]
+        sheets_with_labels = [
+            {
+                "sheetName": "Widgets",
+                "namespaceHeader": "section",
+                "keyHeader": "questionName",
+            },
+            {"sheetName": "Labels", "namespaceHeader": "namespace", "keyHeader": "key"},
+        ]
         LabelsGenerator.generate_labels(
-            excel_file_path, labels_output_folder_path, labels_sheet_names=sheet_names
+            excel_file_path,
+            labels_output_folder_path,
+            sheets_with_labels=sheets_with_labels,
         )
 
     # Call the generate_UI_tests function to generate the common-UI-tests-helpers-template.ts.ts if script enabled